refactor: extract screen composables into separate component files
Break large screen files into smaller, focused component files organized
in per-screen components/ subdirectories. This improves readability,
testability, and reusability of individual UI components.

RecipeListScreen: FolderPickerDialog, SwipeableRecipeCard, RecipeCard,
TagChip, InProgressRecipeCard
RecipeDetailScreen: RecipeContent, RecipeMetadata, ScaleControl,
MeasurementToggle, IngredientSectionContent, InstructionSectionContent
SettingsScreen: ApiKeySection, ModelSelectionSection, DisplaySection,
GoogleDriveSection, AboutSection

Closes #71
