+++ app/src/main/kotlin/com/lionotter/recipes/ui/screens/recipedetail/RecipeDetailScreen.kt
@@ -1,42 +1,17 @@
 package com.lionotter.recipes.ui.screens.recipedetail
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material.icons.filled.Star
 import androidx.compose.material.icons.outlined.StarOutline
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
-import androidx.compose.material3.SuggestionChip
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -46,34 +21,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import android.content.Intent
-import android.net.Uri
-import androidx.core.net.toUri
-import coil.compose.AsyncImage
 import com.lionotter.recipes.R
-import com.lionotter.recipes.domain.model.IngredientSection
-import com.lionotter.recipes.domain.model.IngredientUsageStatus
-import com.lionotter.recipes.domain.model.InstructionIngredientKey
-import com.lionotter.recipes.domain.model.InstructionSection
-import com.lionotter.recipes.domain.model.MeasurementPreference
-import com.lionotter.recipes.domain.model.MeasurementType
-import com.lionotter.recipes.domain.model.Recipe
-import com.lionotter.recipes.domain.model.createInstructionIngredientKey
 import com.lionotter.recipes.ui.components.DeleteConfirmationDialog
 import com.lionotter.recipes.ui.components.RecipeTopAppBar
-import com.lionotter.recipes.util.pluralize
-import com.lionotter.recipes.util.singularize
+import com.lionotter.recipes.ui.screens.recipedetail.components.RecipeContent
 
 @Composable
 fun RecipeDetailScreen(
@@ -171,580 +126,6 @@ fun RecipeDetailScreen(
                 onToggleHighlightedInstruction = viewModel::toggleHighlightedInstructionStep,
                 modifier = Modifier.padding(paddingValues)
             )
-        }
-    }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun RecipeContent(
-    recipe: Recipe,
-    scale: Double,
-    onScaleIncrement: () -> Unit,
-    onScaleDecrement: () -> Unit,
-    measurementPreference: MeasurementPreference,
-    onMeasurementPreferenceChange: (MeasurementPreference) -> Unit,
-    showMeasurementToggle: Boolean,
-    availableMeasurementTypes: Set<MeasurementType>,
-    usedInstructionIngredients: Set<InstructionIngredientKey>,
-    globalIngredientUsage: Map<String, IngredientUsageStatus>,
-    onToggleInstructionIngredient: (Int, Int, Int) -> Unit,
-    highlightedInstructionStep: HighlightedInstructionStep?,
-    onToggleHighlightedInstruction: (Int, Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .verticalScroll(rememberScrollState())
-    ) {
-        // Hero image
-        recipe.imageUrl?.let { imageUrl ->
-            AsyncImage(
-                model = imageUrl,
-                contentDescription = recipe.name,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(16f / 9f)
-                    .clip(MaterialTheme.shapes.medium),
-                contentScale = ContentScale.Crop
-            )
-        }
-
-        Column(modifier = Modifier.padding(16.dp)) {
-            // Tags
-            if (recipe.tags.isNotEmpty()) {
-                FlowRow(
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                    verticalArrangement = Arrangement.spacedBy(4.dp)
-                ) {
-                    recipe.tags.forEach { tag ->
-                        SuggestionChip(
-                            onClick = { },
-                            label = { Text(tag) }
-                        )
-                    }
-                }
-                Spacer(modifier = Modifier.height(16.dp))
-            }
-
-            // Time and servings info
-            RecipeMetadata(recipe = recipe, scale = scale)
-
-            // Story
-            recipe.story?.let { story ->
-                Spacer(modifier = Modifier.height(16.dp))
-                Card(
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.surfaceVariant
-                    )
-                ) {
-                    Text(
-                        text = story,
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontStyle = FontStyle.Italic,
-                        modifier = Modifier.padding(16.dp)
-                    )
-                }
-            }
-
-            // Scale controls
-            Spacer(modifier = Modifier.height(24.dp))
-            ScaleControl(
-                scale = scale,
-                onIncrement = onScaleIncrement,
-                onDecrement = onScaleDecrement
-            )
-
-            // Measurement toggle (only shown if recipe has multiple measurement types)
-            if (showMeasurementToggle) {
-                Spacer(modifier = Modifier.height(16.dp))
-                MeasurementToggle(
-                    selectedPreference = measurementPreference,
-                    onPreferenceChange = onMeasurementPreferenceChange,
-                    availableMeasurementTypes = availableMeasurementTypes
-                )
-            }
-
-            // Ingredients
-            Spacer(modifier = Modifier.height(24.dp))
-            Text(
-                text = stringResource(R.string.ingredients),
-                style = MaterialTheme.typography.titleLarge,
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.height(12.dp))
-            recipe.ingredientSections.forEach { section ->
-                IngredientSectionContent(
-                    section = section,
-                    scale = scale,
-                    measurementPreference = measurementPreference,
-                    globalIngredientUsage = globalIngredientUsage
-                )
-            }
-
-            // Instructions
-            Spacer(modifier = Modifier.height(24.dp))
-            Text(
-                text = stringResource(R.string.instructions),
-                style = MaterialTheme.typography.titleLarge,
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.height(12.dp))
-            recipe.instructionSections.forEachIndexed { sectionIndex, section ->
-                InstructionSectionContent(
-                    section = section,
-                    sectionIndex = sectionIndex,
-                    scale = scale,
-                    measurementPreference = measurementPreference,
-                    usedInstructionIngredients = usedInstructionIngredients,
-                    onToggleIngredient = onToggleInstructionIngredient,
-                    highlightedInstructionStep = highlightedInstructionStep,
-                    onToggleHighlightedInstruction = onToggleHighlightedInstruction
-                )
-            }
-
-            // Source
-            recipe.sourceUrl?.let { url ->
-                val context = LocalContext.current
-                Spacer(modifier = Modifier.height(24.dp))
-                HorizontalDivider()
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(
-                    text = stringResource(R.string.source_prefix, url),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.clickable {
-                        val intent = Intent(Intent.ACTION_VIEW, url.toUri())
-                        context.startActivity(intent)
-                    }
-                )
-            }
-
-            Spacer(modifier = Modifier.height(32.dp))
-        }
-    }
-}
-
-@Composable
-private fun RecipeMetadata(recipe: Recipe, scale: Double) {
-    val items = buildList {
-        recipe.prepTime?.let { add(stringResource(R.string.prep_time, it)) }
-        recipe.cookTime?.let { add(stringResource(R.string.cook_time, it)) }
-        recipe.totalTime?.let { add(stringResource(R.string.total_time, it)) }
-        recipe.servings?.let {
-            val scaled = (it * scale).let { s ->
-                if (s == s.toLong().toDouble()) s.toLong().toString()
-                else "%.1f".format(s)
-            }
-            add(stringResource(R.string.servings_label, scaled))
-        }
-    }
-
-    if (items.isNotEmpty()) {
-        Card(
-            colors = CardDefaults.cardColors(
-                containerColor = MaterialTheme.colorScheme.primaryContainer
-            )
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(12.dp),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                items.forEach { item ->
-                    Text(
-                        text = item,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onPrimaryContainer
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ScaleControl(
-    scale: Double,
-    onIncrement: () -> Unit,
-    onDecrement: () -> Unit
-) {
-    Card(
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.secondaryContainer
-        )
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(8.dp),
-            horizontalArrangement = Arrangement.Center,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = stringResource(R.string.scale_label),
-                style = MaterialTheme.typography.titleSmall,
-                color = MaterialTheme.colorScheme.onSecondaryContainer
-            )
-            Spacer(modifier = Modifier.width(16.dp))
-            IconButton(onClick = onDecrement) {
-                Icon(
-                    imageVector = Icons.Default.Remove,
-                    contentDescription = stringResource(R.string.decrease_scale)
-                )
-            }
-            Text(
-                text = "${scale}x",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onSecondaryContainer
-            )
-            IconButton(onClick = onIncrement) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(R.string.increase_scale)
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun MeasurementToggle(
-    selectedPreference: MeasurementPreference,
-    onPreferenceChange: (MeasurementPreference) -> Unit,
-    availableMeasurementTypes: Set<MeasurementType>
-) {
-    // Determine which options to show based on available measurement types
-    val hasVolume = MeasurementType.VOLUME in availableMeasurementTypes
-    val hasWeight = MeasurementType.WEIGHT in availableMeasurementTypes
-
-    // Build the list of options to display
-    val options = buildList {
-        add(MeasurementPreference.ORIGINAL to stringResource(R.string.original))
-        if (hasVolume) add(MeasurementPreference.VOLUME to stringResource(R.string.volume))
-        if (hasWeight) add(MeasurementPreference.WEIGHT to stringResource(R.string.weight))
-    }
-
-    // Only show toggle if there are at least 2 options (Original plus at least one conversion)
-    if (options.size < 2) return
-
-    Card(
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        )
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = stringResource(R.string.units),
-                style = MaterialTheme.typography.titleSmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(bottom = 8.dp)
-            )
-            SingleChoiceSegmentedButtonRow(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                options.forEachIndexed { index, (preference, label) ->
-                    SegmentedButton(
-                        selected = selectedPreference == preference,
-                        onClick = { onPreferenceChange(preference) },
-                        shape = SegmentedButtonDefaults.itemShape(
-                            index = index,
-                            count = options.size
-                        )
-                    ) {
-                        Text(label)
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun IngredientSectionContent(
-    section: IngredientSection,
-    scale: Double,
-    measurementPreference: MeasurementPreference,
-    globalIngredientUsage: Map<String, IngredientUsageStatus>
-) {
-    Column {
-        section.name?.let { name ->
-            Text(
-                text = name,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-                color = MaterialTheme.colorScheme.primary,
-                modifier = Modifier.padding(vertical = 8.dp)
-            )
-        }
-
-        section.ingredients.forEach { ingredient ->
-            val usage = globalIngredientUsage[ingredient.name.lowercase()]
-            val isFullyUsed = usage?.isFullyUsed == true
-            val hasPartialUsage = usage != null && usage.usedAmount > 0 && !isFullyUsed
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 4.dp)
-            ) {
-                Text(
-                    text = "•",
-                    style = MaterialTheme.typography.bodyLarge,
-                    color = if (isFullyUsed) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.onSurface,
-                    modifier = Modifier.padding(end = 8.dp)
-                )
-                Column {
-                    Text(
-                        text = ingredient.format(scale, measurementPreference),
-                        style = MaterialTheme.typography.bodyLarge,
-                        textDecoration = if (isFullyUsed) TextDecoration.LineThrough else TextDecoration.None,
-                        color = if (isFullyUsed) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.onSurface
-                    )
-                    // Show remaining amount if partially used
-                    if (hasPartialUsage && usage?.remainingAmount != null && usage.remainingAmount > 0) {
-                        Text(
-                            text = formatRemainingAmount(usage.remainingAmount, usage.unit),
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.secondary
-                        )
-                    }
-                }
-            }
-
-            // Display alternates
-            ingredient.alternates.forEach { alternate ->
-                val altUsage = globalIngredientUsage[alternate.name.lowercase()]
-                val altIsFullyUsed = altUsage?.isFullyUsed == true
-                val altHasPartialUsage = altUsage != null && altUsage.usedAmount > 0 && !altIsFullyUsed
-
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 2.dp, horizontal = 24.dp)
-                ) {
-                    Text(
-                        text = stringResource(R.string.or),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.outline,
-                        modifier = Modifier.padding(end = 8.dp)
-                    )
-                    Column {
-                        Text(
-                            text = alternate.format(scale, measurementPreference),
-                            style = MaterialTheme.typography.bodyMedium,
-                            textDecoration = if (altIsFullyUsed) TextDecoration.LineThrough else TextDecoration.None,
-                            color = if (altIsFullyUsed) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.onSurfaceVariant
-                        )
-                        // Show remaining amount if partially used
-                        if (altHasPartialUsage && altUsage?.remainingAmount != null && altUsage.remainingAmount > 0) {
-                            Text(
-                                text = formatRemainingAmount(altUsage.remainingAmount, altUsage.unit),
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.secondary
-                            )
-                        }
-                    }
-                }
-            }
-        }
-
-        if (section.name != null) {
-            Spacer(modifier = Modifier.height(8.dp))
-        }
-    }
-}
-
-/**
- * Formats the remaining amount for display (e.g., "1/2 cup left").
- */
-private fun formatRemainingAmount(amount: Double, unit: String?): String {
-    val formattedAmount = formatQuantity(amount)
-    return if (unit != null) {
-        val count = if (amount > 1.0) 2 else 1
-        val pluralizedUnit = unit.singularize().pluralize(count)
-        "$formattedAmount $pluralizedUnit left"
-    } else {
-        "$formattedAmount left"
-    }
-}
-
-private fun formatQuantity(qty: Double): String {
-    return if (qty == qty.toLong().toDouble()) {
-        qty.toLong().toString()
-    } else {
-        val fractions = mapOf(
-            0.25 to "1/4",
-            0.33 to "1/3",
-            0.5 to "1/2",
-            0.66 to "2/3",
-            0.75 to "3/4"
-        )
-        val whole = qty.toLong()
-        val decimal = qty - whole
-
-        val fraction = fractions.entries.minByOrNull {
-            kotlin.math.abs(it.key - decimal)
-        }?.takeIf {
-            kotlin.math.abs(it.key - decimal) < 0.05
-        }?.value
-
-        when {
-            fraction != null && whole > 0 -> "$whole $fraction"
-            fraction != null -> fraction
-            else -> "%.2f".format(qty).trimEnd('0').trimEnd('.')
-        }
-    }
-}
-
-@Composable
-private fun InstructionSectionContent(
-    section: InstructionSection,
-    sectionIndex: Int,
-    scale: Double,
-    measurementPreference: MeasurementPreference,
-    usedInstructionIngredients: Set<InstructionIngredientKey>,
-    onToggleIngredient: (Int, Int, Int) -> Unit,
-    highlightedInstructionStep: HighlightedInstructionStep?,
-    onToggleHighlightedInstruction: (Int, Int) -> Unit
-) {
-    Column {
-        section.name?.let { name ->
-            Text(
-                text = name,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-                color = MaterialTheme.colorScheme.primary,
-                modifier = Modifier.padding(vertical = 8.dp)
-            )
-        }
-
-        section.steps.forEachIndexed { stepIndex, step ->
-            val isHighlighted = highlightedInstructionStep?.sectionIndex == sectionIndex &&
-                                highlightedInstructionStep?.stepIndex == stepIndex
-
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 8.dp)
-                    .clickable { onToggleHighlightedInstruction(sectionIndex, stepIndex) }
-                    .then(
-                        if (isHighlighted) {
-                            Modifier
-                                .clip(MaterialTheme.shapes.medium)
-                                .background(MaterialTheme.colorScheme.tertiaryContainer)
-                        } else {
-                            Modifier
-                        }
-                    )
-                    .padding(horizontal = 8.dp, vertical = 8.dp)
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    Card(
-                        colors = CardDefaults.cardColors(
-                            containerColor = if (isHighlighted)
-                                MaterialTheme.colorScheme.secondaryContainer
-                            else
-                                MaterialTheme.colorScheme.primary
-                        ),
-                        modifier = Modifier.padding(end = 12.dp)
-                    ) {
-                        Text(
-                            text = "${step.stepNumber}",
-                            style = MaterialTheme.typography.labelLarge,
-                            color = if (isHighlighted)
-                                MaterialTheme.colorScheme.onSecondaryContainer
-                            else
-                                MaterialTheme.colorScheme.onPrimary,
-                            modifier = Modifier.padding(horizontal = 10.dp, vertical = 4.dp)
-                        )
-                    }
-                    Text(
-                        text = step.instruction,
-                        style = MaterialTheme.typography.bodyLarge,
-                        modifier = Modifier.weight(1f),
-                        color = if (isHighlighted)
-                            MaterialTheme.colorScheme.onSurface
-                        else
-                            MaterialTheme.colorScheme.onSurface
-                    )
-                }
-
-                // Display step-level ingredients if present
-                if (step.ingredients.isNotEmpty()) {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(start = 40.dp, top = 8.dp)
-                    ) {
-                        step.ingredients.forEachIndexed { ingredientIndex, ingredient ->
-                            val ingredientKey = createInstructionIngredientKey(sectionIndex, stepIndex, ingredientIndex)
-                            val isUsed = ingredientKey in usedInstructionIngredients
-
-                            // Main ingredient row - clickable
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .clickable { onToggleIngredient(sectionIndex, stepIndex, ingredientIndex) }
-                                    .padding(vertical = 4.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Text(
-                                    text = if (isUsed) "✓" else "•",
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    color = if (isUsed) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.secondary,
-                                    modifier = Modifier.padding(end = 8.dp)
-                                )
-                                Text(
-                                    text = ingredient.format(scale, measurementPreference),
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    textDecoration = if (isUsed) TextDecoration.LineThrough else TextDecoration.None,
-                                    color = if (isUsed) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.onSurface
-                                )
-                            }
-
-                            // Display alternates for step ingredients - also clickable (same toggle)
-                            ingredient.alternates.forEach { alternate ->
-                                Row(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .clickable { onToggleIngredient(sectionIndex, stepIndex, ingredientIndex) }
-                                        .padding(vertical = 2.dp, horizontal = 24.dp),
-                                    verticalAlignment = Alignment.CenterVertically
-                                ) {
-                                    Text(
-                                        text = stringResource(R.string.or),
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.outline,
-                                        modifier = Modifier.padding(end = 8.dp)
-                                    )
-                                    Text(
-                                        text = alternate.format(scale, measurementPreference),
-                                        style = MaterialTheme.typography.bodySmall,
-                                        textDecoration = if (isUsed) TextDecoration.LineThrough else TextDecoration.None,
-                                        color = if (isUsed) MaterialTheme.colorScheme.outline else MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        if (section.name != null) {
-            Spacer(modifier = Modifier.height(8.dp))
         }
     }
 }
