+++ app/src/main/kotlin/com/lionotter/recipes/ui/screens/settings/components/ApiKeySection.kt
@@ -1,0 +1,170 @@
+package com.lionotter.recipes.ui.screens.settings.components
+
+import android.content.Intent
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Check
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Visibility
+import androidx.compose.material.icons.filled.VisibilityOff
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.core.net.toUri
+import com.lionotter.recipes.R
+import com.lionotter.recipes.ui.screens.settings.SettingsViewModel
+
+@Composable
+fun ApiKeySection(
+    currentApiKey: String?,
+    apiKeyInput: String,
+    onApiKeyInputChange: (String) -> Unit,
+    onSaveClick: () -> Unit,
+    onClearClick: () -> Unit,
+    saveState: SettingsViewModel.SaveState,
+    onResetSaveState: () -> Unit
+) {
+    var passwordVisible by remember { mutableStateOf(false) }
+    val context = LocalContext.current
+
+    Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
+        Text(
+            text = stringResource(R.string.anthropic_api_key),
+            style = MaterialTheme.typography.titleMedium
+        )
+
+        Text(
+            text = stringResource(R.string.api_key_storage_info),
+            style = MaterialTheme.typography.bodySmall,
+            color = MaterialTheme.colorScheme.onSurfaceVariant
+        )
+
+        if (currentApiKey != null) {
+            Card(
+                colors = CardDefaults.cardColors(
+                    containerColor = MaterialTheme.colorScheme.primaryContainer
+                )
+            ) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Column(modifier = Modifier.weight(1f)) {
+                        Row(verticalAlignment = Alignment.CenterVertically) {
+                            Icon(
+                                imageVector = Icons.Default.Check,
+                                contentDescription = stringResource(R.string.api_key_configured),
+                                tint = MaterialTheme.colorScheme.primary
+                            )
+                            Text(
+                                text = " " + stringResource(R.string.api_key_configured_label),
+                                style = MaterialTheme.typography.bodyMedium,
+                                color = MaterialTheme.colorScheme.onPrimaryContainer
+                            )
+                        }
+                        Text(
+                            text = stringResource(R.string.api_key_masked, currentApiKey.takeLast(4)),
+                            style = MaterialTheme.typography.bodySmall,
+                            color = MaterialTheme.colorScheme.onPrimaryContainer.copy(alpha = 0.7f)
+                        )
+                    }
+                    IconButton(onClick = onClearClick) {
+                        Icon(
+                            imageVector = Icons.Default.Delete,
+                            contentDescription = stringResource(R.string.remove_api_key),
+                            tint = MaterialTheme.colorScheme.error
+                        )
+                    }
+                }
+            }
+        } else {
+            OutlinedTextField(
+                value = apiKeyInput,
+                onValueChange = onApiKeyInputChange,
+                modifier = Modifier.fillMaxWidth(),
+                label = { Text(stringResource(R.string.api_key)) },
+                placeholder = { Text(stringResource(R.string.api_key_placeholder)) },
+                singleLine = true,
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
+                visualTransformation = if (passwordVisible) {
+                    VisualTransformation.None
+                } else {
+                    PasswordVisualTransformation()
+                },
+                trailingIcon = {
+                    IconButton(onClick = { passwordVisible = !passwordVisible }) {
+                        Icon(
+                            imageVector = if (passwordVisible) {
+                                Icons.Default.VisibilityOff
+                            } else {
+                                Icons.Default.Visibility
+                            },
+                            contentDescription = if (passwordVisible) stringResource(R.string.hide) else stringResource(R.string.show)
+                        )
+                    }
+                },
+                isError = saveState is SettingsViewModel.SaveState.Error
+            )
+
+            if (saveState is SettingsViewModel.SaveState.Error) {
+                Text(
+                    text = saveState.message,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.error
+                )
+            }
+
+            if (saveState is SettingsViewModel.SaveState.Success) {
+                LaunchedEffect(Unit) {
+                    onResetSaveState()
+                }
+            }
+
+            Button(
+                onClick = onSaveClick,
+                modifier = Modifier.fillMaxWidth(),
+                enabled = apiKeyInput.isNotBlank()
+            ) {
+                Text(stringResource(R.string.save_api_key))
+            }
+        }
+
+        OutlinedButton(
+            onClick = {
+                val intent = Intent(Intent.ACTION_VIEW, "https://platform.claude.com/settings/keys".toUri())
+                context.startActivity(intent)
+            },
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text(stringResource(R.string.get_api_key))
+        }
+    }
+}
